Updates... * Move constants to new dedicated file * Move remaining transform functions to transform file * Provide top-level exports of minifyCss and prefixCss * Babel dependencies
