+++ package.json
@@ -7,21 +7,20 @@
     "autoprefixer": "^6.4.0",
     "blob-polyfill": "^1.0.20150320",
     "postcss": "^5.1.2",
-    "react": ">=15.0.0"
+    "react": ">=15.3.1"
   },
   "description": "Write scoped, autoprefixed styles declaratively in React",
   "devDependencies": {
     "babel": "6.5.2",
-    "babel-cli": "6.11.4",
+    "babel-cli": "6.14.0",
     "babel-eslint": "6.1.2",
     "babel-loader": "6.2.5",
     "babel-plugin-add-module-exports": "0.2.1",
     "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "6.13.2",
+    "babel-preset-es2015": "6.14.0",
     "babel-preset-es2015-loose": "^7.0.0",
     "babel-preset-react": "6.11.1",
     "babel-preset-stage-0": "6.5.0",
-    "babel-register": "^6.11.6",
     "eslint": "3.3.1",
     "eslint-friendly-formatter": "2.0.6",
     "eslint-loader": "1.5.0",
@@ -54,5 +53,5 @@
     "prepublish": "in-publish && npm run compile-for-publish || in-install",
     "transpile": "node ./node_modules/babel-cli/lib/babel/index.js -d lib/ src/"
   },
-  "version": "1.2.0"
+  "version": "1.2.1"
 }
